Initial Electric Vehicle tab version
Initial version of a EV-data dedicated tab (Electric Vehicle). Combines
all the Wheellog CVS data in a single view, but contains the last
loaded file only.

The display of individual EV parameters can be turned on/off with a
context menu. This selection is preserved by using QSettings.

The "Speed" tab now displays the GPS speed, the EV speed, power and
temperature are displayed in "Electric Vehicle" tab only.

TODO:
The parameters are combined w/o re-scaling, thus the ones with smaller
values (Current) are difficult to see when displayed together with the
ones with lager values (Total Distance). The unnecessary parameters must
be hidden as a workaround.