Zoom the graphs with mouse wheel
Mouse wheel zoom to the GraphView. This is as simple as possible
implementation, horizontal zoom only.
A scroll bar is used to pan the data. The whole scene is panned, thus
the axes might not be seen.
When zooming, the view is re-scrolled, in order the data under the
cursor not to move. However, this is still not very precise.