Zoomed graph view artifacts workaround
Workaround for the nasty ghost graph item artifacts, when scrolling
zoomed (> 2x) view that contains negative values. The artifacts are
possibly caused by the drawing outside of the visible area.

The workaround reorders the drawing of the horizontal grid lines before
the vertical ones. Actually, a single horizontal line to be pre-drawn
along the entire view,  is enough to fix these artifacts.