Fix spaghetti bug in the "Electric Vehicle" tab
In Graph -> Distance mode, when "Eliminate GPS outliers" is checked
(Settings -> Options... -> Data), the "Electric Vehicle" tab often shows
"spaghetty" graphs. This is because the points selected as "outliers"
were not ignored by Track::evScalar(). Since, the distances are
adjusted, but not the "outliers", the combined distances were not always
in increasing order.